don't complain if you can't find the file

+++ st/templatetags/library.py
@@ -20,16 +20,19 @@ def thumbnail(file, size='200x200'):
     miniature_filename = os.path.join(MEDIA_ROOT, miniature)
     miniature_url = os.path.join(MEDIA_URL, miniature)
     filename = os.path.join(MEDIA_ROOT, file)
-    # if image has been modified, remove old thumbnail
-    if os.path.exists(miniature_filename) and os.path.getmtime(filename)>os.path.getmtime(miniature_filename):
-        os.unlink(miniature_filename)
-    # if the image wasn't already resized, resize it
-    if not os.path.exists(miniature_filename):
-        print '>>> debug: resizing the image to the format %s!' % size
-        image = Image.open(filename)
-        image.thumbnail([x, y]) # generate a 200x200 thumbnail
-        image.save(miniature_filename, image.format)
-    return miniature_url
+    if os.path.exists(filename):
+        # if image has been modified, remove old thumbnail
+        if os.path.exists(miniature_filename) and os.path.getmtime(filename)>os.path.getmtime(miniature_filename):
+            os.unlink(miniature_filename)
+        # if the image wasn't already resized, resize it
+        if not os.path.exists(miniature_filename):
+            print '>>> debug: resizing the image to the format %s!' % size
+            image = Image.open(filename)
+            image.thumbnail([x, y]) # generate a 200x200 thumbnail
+            image.save(miniature_filename, image.format)
+        return miniature_url
+    else:
+        return file
 
 @register.filter
 def renderbbcode(context):
